+++ src/components/DropdownMenu/index.ts
@@ -1,2 +1,5 @@
 export { default } from "./DropdownMenu";
-export type { Props as DropdownMenuProps } from "./DropdownMenu.types";
+export type {
+	Props as DropdownMenuProps,
+	Instance as DropdownMenuInstance,
+} from "./DropdownMenu.types";